cr. moved instruction button and added scss files
