+++ src/components/Lesson/Lesson.scss
@@ -18,7 +18,3 @@
   // responsive font
   @include responsiveFontSizes;
 }
-
-.buttonMargin {
-  margin: 10px 10px 0 0;
-}
