+++ src/components/Lesson/ReadmeButton.scss
@@ -1,0 +1,3 @@
+.buttonMargin {
+  margin: 10px 10px 0 0;
+}
