+++ src/selectors/frontmatter.js
@@ -10,11 +10,6 @@ export const getTitle = (state, params) => getLessonFrontmatter(state, params).t
 
 export const getLevel = (state, params) => getLessonFrontmatter(state, params).level || 0;
 
-export const getCourseIcon = (state, {course}) =>
-  course ?
-    state.context.iconContext('./' + course + '/logo-black.png') :
-    null;
-
 export const getAuthorName = (state, params) => getLessonFrontmatter(state, params).author || '';
 
 export const getTranslatorName = (state, params) => getLessonFrontmatter(state, params).translator || '';
